requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
mysql-connector-python==8.2.0
pandas==2.1.1
openpyxl==3.1.2
Werkzeug==3.0.1
